feat: use @vue-storefront/nuxt to setup the sdk
