+++ apps/web/composables/useAdditionalInformation/useAdditionalInformation.ts
@@ -1,6 +1,6 @@
 import type { AdditionalInformationParams } from '@plentymarkets/shop-api';
 import { toRefs } from '@vueuse/shared';
-import { useSdk } from '~/sdk';
+
 import {
   DoAdditionalInformation,
   DoAdditionalInformationReturn,
