+++ apps/web/composables/useAddress/useAddress.ts
@@ -2,7 +2,7 @@ import { Address, AddressType } from '@plentymarkets/shop-api';
 import { userAddressGetters } from '@plentymarkets/shop-sdk';
 import { toRefs } from '@vueuse/shared';
 import { DeleteAddress, SetDefault } from '~/composables/useAddress/types';
-import { useSdk } from '~/sdk';
+
 import { UseAddressReturn, GetAddresses, SaveAddress, UseAddressMethodsState } from './types';
 
 /**
