+++ apps/web/composables/useCustomer/useCustomer.ts
@@ -10,7 +10,6 @@ import type {
   Logout,
   ChangePassword,
 } from '~/composables/useCustomer/types';
-import { useSdk } from '~/sdk';
 
 /**
  * @description Composable managing Customer data
