+++ apps/web/composables/useCustomerOrders/useCustomerOrders.ts
@@ -4,7 +4,7 @@ import type {
   UseCustomerOrdersState,
   FetchCustomerOrders,
 } from '~/composables/useCustomerOrders/types';
-import { useSdk } from '~/sdk';
+
 import { UseUserOrderSearchParams } from '@plentymarkets/shop-api';
 
 /**
