+++ apps/web/composables/useLegalInformation/useLegalInformation.ts
@@ -1,6 +1,6 @@
 import type { LegalInformationResponse, LegalTextsParams } from '@plentymarkets/shop-api';
 import { toRefs } from '@vueuse/shared';
-import { useSdk } from '~/sdk';
+
 import type { UseLegalInformationState, UseLegalInformationMethodsReturn, GetLegalInformation } from './types';
 
 /**
