+++ apps/web/composables/useOrderDocument/useOrderDocument.ts
@@ -1,6 +1,6 @@
 import { OrderDocument } from '@plentymarkets/shop-api';
 import { toRefs } from '@vueuse/shared';
-import { useSdk } from '~/sdk';
+
 import type { UseOrderDocumentState, UseOrderDocumentMethodsReturn } from './types';
 import { DownloadFile, GetDocument } from './types';
