+++ apps/web/composables/usePaymentMethods/usePaymentMethods.ts
@@ -1,6 +1,6 @@
 import type { PaymentProviders } from '@plentymarkets/shop-api';
 import { toRefs } from '@vueuse/shared';
-import { useSdk } from '~/sdk';
+
 import { UsePaymentMethodsReturn, UsePaymentMethodsState, FetchPaymentMethods, SavePaymentMethod } from './types';
 
 /**
