+++ apps/web/composables/useProduct/useProduct.ts
@@ -2,7 +2,6 @@ import type { Product, ProductParams } from '@plentymarkets/shop-api';
 import { categoryTreeGetters, productGetters } from '@plentymarkets/shop-sdk';
 import { toRefs } from '@vueuse/shared';
 import type { UseProductReturn, UseProductState, FetchProduct } from '~/composables/useProduct/types';
-import { useSdk } from '~/sdk';
 
 /**
  * @description Composable managing product data
