+++ apps/web/composables/useProductReviewAverage/useProductReviewAverage.ts
@@ -1,6 +1,6 @@
 import { ReviewAverage } from '@plentymarkets/shop-api';
 import { toRefs } from '@vueuse/shared';
-import { useSdk } from '~/sdk';
+
 import { FetchProductReviewAverage, UseProductReviewAverage, UseProductReviewAverageState } from './types';
 
 /**
