+++ apps/web/composables/useProductReviews/useProductReviews.ts
@@ -1,6 +1,6 @@
 import type { Review } from '@plentymarkets/shop-api';
 import { toRefs } from '@vueuse/shared';
-import { useSdk } from '~/sdk';
+
 import type { FetchProductReviews, UseProductReviews, UseProductReviewsState } from './types';
 
 /**
