+++ apps/web/composables/useProducts/useProducts.ts
@@ -2,7 +2,6 @@ import { FacetSearchCriteria, Product } from '@plentymarkets/shop-api';
 import type { Facet } from '@plentymarkets/shop-api';
 import { defaults, SelectVariation } from '~/composables';
 import { FetchProducts, UseProductsReturn, UseProductsState } from '~/composables/useProducts/types';
-import { useSdk } from '~/sdk';
 
 /**
  * @description Composable for managing products.
