+++ apps/web/composables/useSearch/useSearch.ts
@@ -1,7 +1,6 @@
 import { ItemSearchParams, ItemSearchResult } from '@plentymarkets/shop-api';
 import { defaults } from '~/composables';
 import { UseSearchReturn, UseSearchState, GetSearch } from '~/composables/useSearch/types';
-import { useSdk } from '~/sdk';
 
 /**
  * @description Composable for managing products search.
