+++ apps/web/package.json
@@ -23,6 +23,7 @@
     "test:cypress-dev": "cypress open --e2e"
   },
   "dependencies": {
+    "@vue-storefront/nuxt": "~3.0.0",
     "@vue-storefront/sdk": "^1.0.1",
     "@vueuse/core": "^10.3.0",
     "lodash-es": "^4.17.21",
